+++ mopidy_bandcamp/__init__.py
@@ -20,11 +20,12 @@ class Extension(ext.Extension):
 
     def get_config_schema(self):
         schema = super(Extension, self).get_config_schema()
-        schema["art_url_as_comment"] = config.Boolean(optional=True)
         schema["discover_pages"] = config.Integer(optional=True)
+        schema["collection_items"] = config.Integer(optional=True)
         schema["discover_genres"] = config.List(optional=True)
         schema["discover_tags"] = config.List(optional=True)
         schema["image_sizes"] = config.List(optional=True)
+        schema["identity"] = config.String(optional=True)
         return schema
 
     def get_command(self):